adds changes to controllers
