+++ controllers/contacts/getAll.js
@@ -1,0 +1,9 @@
+const {Contact} = require("../../models/contact");
+const { ctrlWrapper } = require("../../helpers");
+
+const getAll = async (req, res) => {
+    const result = await Contact.find({});
+    res.json(result);
+}
+
+module.exports = {getAll: ctrlWrapper(getAll)}
