+++ controllers/contacts/updateFavorite.js
@@ -1,0 +1,13 @@
+const {Contact} = require("../../models/contact");
+const { HttpError, ctrlWrapper } = require("../../helpers");
+
+const updateFavorite = async (req, res) => {
+    const { id } = req.params;
+    const result = await Contact.findByIdAndUpdate(id, req.body, {new: true});
+    if (!result) {
+        throw HttpError(404, "Not found");
+    }
+    res.json(result);
+}
+
+module.exports = {updateFavorite: ctrlWrapper(updateFavorite),}
